+++ lib/fab/src/tree/eval.c
@@ -45,6 +45,7 @@ float eval_f(MathTree* tree, const float x, const float y, const float z)
                 case OP_NEG:    node->results.f = neg_f(A); break;
                 case OP_EXP:    node->results.f = exp_f(A); break;
                 case OP_FLOOR:    node->results.f = floor_f(A); break;
+                case OP_LOG:    node->results.f = log_f(A); break;
 
                 case OP_X:      node->results.f = X_f(x); break;
                 case OP_Y:      node->results.f = Y_f(y); break;
@@ -100,6 +101,7 @@ Interval eval_i(MathTree* tree, const Interval X,
                 case OP_NEG:    node->results.i = neg_i(A); break;
                 case OP_EXP:    node->results.i = exp_i(A); break;
                 case OP_FLOOR:    node->results.i = floor_i(A); break;
+                case OP_LOG:    node->results.i = log_i(A); break;
 
                 case OP_CONST:  break;
                 case OP_X:      node->results.i = X_i(X); break;
@@ -153,6 +155,7 @@ float* eval_r(MathTree* tree, const Region r)
                 case OP_NEG:    neg_r(A, R, c); break;
                 case OP_EXP:    exp_r(A, R, c); break;
                 case OP_FLOOR:    floor_r(A, R, c); break;
+                case OP_LOG:    log_r(A, R, c); break;
 
                 case OP_CONST:  break;
                 case OP_X:      X_r(r.X, R, c); break;
@@ -208,6 +211,7 @@ derivative* eval_g(MathTree* tree, const Region r)
                 case OP_NEG:    neg_g(A, R, c); break;
                 case OP_EXP:    exp_g(A, R, c); break;
                 case OP_FLOOR:    floor_g(A, R, c); break;
+                case OP_LOG:    log_g(A, R, c); break;
 
                 case OP_CONST:  break;
                 case OP_X:      X_g(r.X, R, c); break;